requirements.txt

google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
fastapi
Authlib
itsdangerous
psycopg2-binary
python-jose
python-multipart
uvicorn
passlib
sqlalchemy
dotenv
httpx
bcrypt
pandas
google-auth